==> 0_get_labels.ipynb <==
#### Get the labels for the training data

import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
import os
from keras.utils import np_utils

# Define paths
base_path = os.path.abspath("./") # Your root directory
dataset_path = os.path.join(base_path, "dataset") # Your dataset folder

os.listdir(dataset_path)

dt = pd.read_csv(os.path.join(dataset_path, '8ba42c65-534a-4cdf-b983-e7fcaaaa3002_TrainingSet_metaData.csv'))
dt.head()

svr_scores = dt['SVR_Severity'].values
len(svr_scores)

svr_scores

svr_score_binary = []

for s in svr_scores:
    if s == "LOW":
        svr_score_binary.append(0)
    if s == "HIGH":
        svr_score_binary.append(1)
        
len(svr_score_binary)

svr_score_binary = np.array(svr_score_binary)
svr_score_binary.shape

svr_score_binary

type(svr_score_binary)

num_classes = 2

y_train = np_utils.to_categorical(svr_score_binary, num_classes)
y_train.shape

y_train

np.save(dataset_path + "/y_train_clef.npy", y_train)
