==> 1_data_process_clef19.ipynb <==
### Construct numpy array for whole training dataset using the uniformizing techniques

Had to be run three times for each technique. (Refactor!)

* SSS
* ESS
* SIZ

# Run this cell once
#!pip install SimpleITK # for NIFTI file conversion to numpy array
#!pip install tqdm # for seeing progress

#### ATTENTION

Put all NIFTI files from TrainingSet_2_of_2 to TrainingSet_1_of_2, you should get 218 NIFTI files

import SimpleITK as sitk
import os
from tqdm import tqdm
import numpy as np 
import pandas as pd
import os
import scipy.ndimage
import matplotlib.pyplot as plt
from numpy.random import rand
from skimage import measure, morphology
import cv2
from scipy.ndimage import zoom
%matplotlib inline


np.random.seed(1)

# Helpers functions

def create_directory(directory):
    '''
    Creates a new folder in the specified directory if the folder doesn't exist.
    INPUT
        directory: Folder to be created, called as "folder/".
    OUTPUT
        New folder in the current directory.
    '''
    if not os.path.exists(directory):
        os.makedirs(directory)


# Define paths
base_path = os.path.abspath("./") # Your root directory
dataset_path = os.path.join(base_path, "dataset") # Your dataset folder
model_path = os.path.join(base_path, "models")
log_path = os.path.join(base_path, "logs")

# Raw data path
INPUT_FOLDER = 'TrainingSet_1_of_2/'  
training_data_raw = os.path.join(dataset_path, INPUT_FOLDER)

# Check inside target folder
print(os.listdir(dataset_path))
print(training_data_raw)

patients = sorted(os.listdir(training_data_raw))
print(patients[:10])

y_train = np.load("{}/y_train_clef.npy".format(dataset_path))
print(y_train.shape)

# Resize 2D slices
w, h = 128, 128
def rs_img(img):
    '''W and H is 128 now
    '''
    img = np.transpose(img)
    flatten = [cv2.resize(img[:,:,i], (w, h), interpolation=cv2.INTER_CUBIC) for i in range(img.shape[-1])]
    img = np.array(np.dstack(flatten)) 
    return img


# Subset slice selection (SSS)
def change_depth_sss(img):
    
    factor = 16
    img_start = img[:,:,:factor]
    
    mid = int(img.shape[-1]/2)
    img_middle = img[:,:,mid-factor:mid+factor]
    
    img_end = img[:,:,-factor:]
    img = np.concatenate((img_start, img_middle, img_end), axis=2)
    return img


# Even slice selection (ESS)
def change_depth_ess(img):
    
    '''ESS depth size is 32 as min depth size is 47, so we cannot evenly select 64 slices!'''

    # Ignore volumes which have depth less than 64
    target_depth = 64

    #print(img.shape) 

    scan_depth = int(img.shape[-1])
    
    # If depth is not 64, make it 64
    if scan_depth < 64:
        number = 64 - scan_depth
        #print(number)
        # take last image and stack
        extend = [img[:,:,-1] for _ in range(number)]
        extend = np.array(np.dstack(extend))
        #print(extend.shape)
        img = np.concatenate((img, extend), axis=2)
    
    else:
        factor = int(np.floor(scan_depth / target_depth)) # min depth is 32
        #print(factor) # floor and ceil result in different values
        #print(scan_depth, factor)

        flatten = []
        idx = 0

        for i in range(0, scan_depth, factor):
          #print(idx)

            if idx>=scan_depth:
                break
            else:
                flatten.append(img[:,:,i])

        img = np.array(np.dstack(flatten))

    # hardcode from bottom, if not 64 slices
    img = img[:,:,:target_depth]     

    #print("Final shape: ", img.shape)
    
    # hardcode from bottom, if not 32 slices
    img = img[:,:,:target_depth]

    assert img.shape[-1] == target_depth , "Error"
    return img


# Spline interpolated zoom (SIZ)
def change_depth_siz(img):
    desired_depth = 64
    current_depth = img.shape[-1]
    depth = current_depth / desired_depth
    depth_factor = 1 / depth
    img_new = zoom(img, (1, 1, depth_factor), mode='nearest')
    return img_new


#def normalize(image):
    
 #   MIN_BOUND = -1000 # anything below this we are not interested in 
 #   MAX_BOUND = 700   # anything above too

   # image = (image - MIN_BOUND) / (MAX_BOUND - MIN_BOUND)
  #  image[image>1] = 1.
  #  image[image<0] = 0.
  #  return image

    
def normalize(image):
    global MIN_BOUND
    global MAX_BOUND
    
    image = (image - MIN_BOUND) / (MAX_BOUND - MIN_BOUND)
    image[image>1] = 1.
    image[image<0] = 0.
    return image


def zero_center(image):
    image = image - PIXEL_MEAN
    return image


def plot_seq(data, name):
    
    a, b = 3, 20
    data = np.reshape(data, (a, b, 128, 128))
    test_data = data
    r, c = test_data.shape[0], test_data.shape[1]

    cmaps = [['viridis', 'binary'], ['plasma', 'coolwarm'], ['Greens', 'copper']]

    heights = [a[0].shape[0] for a in test_data]
    widths = [a.shape[1] for a in test_data[0]]

    fig_width = 10.  # inches
    fig_height = fig_width * sum(heights) / sum(widths)

    f, axarr = plt.subplots(r,c, figsize=(fig_width, fig_height),
          gridspec_kw={'height_ratios':heights})

    for i in range(r):
        for j in range(c):
            axarr[i, j].imshow(test_data[i][j], cmap='gray')
            axarr[i, j].axis('off')
    
    plt.subplots_adjust(wspace=0, hspace=0, left=0, right=1, bottom=0, top=1)
    plt.savefig('{}/{}.png'.format(log_path, name), dpi=1000)
    plt.show()


# NOT USING THIS NOW
def resize_depth_wise(img3d):
    '''
      Inputs a 3d tensor with uneven depth
      Outputs a 3d tensor with even depth, in this case depth=64
    '''
  
  # patient image 3D
    p = img3d
  # list of 3D slices of p
    p_2d = []

    depth = 2

    n = 0
    c = 0

    for c in range(70):
        img = img3d[:,:,n+depth]
        p_2d.append(img)
        n = n+depth
        c = c+1
        
    p_3d_d64 = np.array(np.dstack(p_2d))
    return p_3d_d64


# For debugging

img = sitk.ReadImage(training_data_raw + patients[0])
img = sitk.GetArrayFromImage(img)
img = rs_img(img)
print("Raw volume size: ", img.shape)

img_ess = change_depth_ess(img)
img_sss = change_depth_sss(img)
img_siz = change_depth_siz(img)

img_sss = np.transpose(img_sss)
img_ess = np.transpose(img_ess)
img_siz = np.transpose(img_siz)

print(img_sss.shape, img_ess.shape, img_siz.shape)


vols = [img_sss[:60], img_ess[:60], img_siz[:60]]
tech = ['sss', 'ess', 'siz']

for v, t in zip(vols, tech):
    plot_seq(v, t)

#img = np.transpose(img)
#plot_seq(img[:120], "raw")

# Load training data


# Get a raw data
# Resize


depths = []
pts = []
#y_train_ess = []
#ct = 0

for ps in tqdm(patients[:]):

    #print("Image Name", ps)
    img = sitk.ReadImage(training_data_raw + ps)
    img = sitk.GetArrayFromImage(img)
    
    # Resize x,y,z
    img = rs_img(img)

    # Get depth of each volume
    d = img.shape[-1]
    depths.append(d)
    
    # What downsampling technique?
    img = change_depth_siz(img)

    # Normalize [0, 1]
    #img = normalize(img)

    pts.append(img)
    img=None
    #y_train_ess.append(y_train[ct])
    #ct+=1

# Convert to numpy array
train_data = np.array(pts)
print(train_data.shape)

# Clear memory
pts = None


print("Min depth:", min(depths))

#y_train_ess = np.array(y_train_ess)
#print(y_train_ess.shape)

#np.save('{}/y_train_clef_ess.npy'.format(dataset_path), y_train_ess)

# Get count of volumes < 64
count = 0

for i in depths:
    if i < 64:
        count+=1

print(count)

######################################################################
# Test data path
#INPUT_FOLDER = 'TestSet/'  
#test_data_raw = os.path.join(dataset_path, INPUT_FOLDER)

#patients_test = sorted(os.listdir(test_data_raw))
#print(len(patients_test))
#print(patients_test[:2])

# Load test data
#pts_test = []

#for ps in tqdm(patients_test[:]):

    #print("Image Name", ps)
#    img = sitk.ReadImage(test_data_raw + ps)
#    img = sitk.GetArrayFromImage(img)
    
    # Resize x,y,z
#    img = rs_img(img)

    # What downsampling technique?
#    img = change_depth_sss(img)

#    pts_test.append(img)
#    img=None

# Convert to numpy array
#test_data = np.array(pts_test)
#test_data.shape
######################################################################

train_data.shape

img = train_data[0]
img.shape

np.min(img), np.max(img)

plt.imshow(img[:,:,25], cmap='gray')

img[:,:,30]

# Get these values for preprocessing

MIN_BOUND = np.min(train_data)
MAX_BOUND = np.max(train_data)
PIXEL_MEAN = np.mean(train_data)

print("Min", MIN_BOUND)
print("Min", MAX_BOUND)
print("Mean", PIXEL_MEAN)

# Normalize
x_train = normalize(train_data)

np.max(x_train), np.min(x_train)

#https://github.com/hasibzunair/uniformizing-3D/blob/master/others/5th_place_submission/preprocessing/normalization_and_zero_center.ipynb

# Zero center data
#x_train_zero = zero_center(x_train)

# After processing

MIN_BOUND = np.min(x_train)
MAX_BOUND = np.max(x_train)
PIXEL_MEAN = np.mean(x_train)

print("Min", MIN_BOUND)
print("Max", MAX_BOUND)
print("Mean", PIXEL_MEAN)

# Save training and test data
#np.save('{}/x_test.npy'.format(dataset_path), x_test)
np.save('{}/x_train_siz_norm.npy'.format(dataset_path), x_train)
